Fix queue_disconnected_test
When running list_keys there are two timeouts.  The PB server timeout
and the list keys timeout.  Since the time this test was written, the
PB server timeout for the `list_keys` call was changed to `infinity`
and instead the timeout is passed to the list keys operation.  Thus
this test will timeout waiting indefinitely for the `list_keys` call
to return.  I've simply removed this assertion as it requires a
different context to be tested.