Allow modify_type/5 to execute the update on not-found.
When a key containing a data-type is not found on fetch, the socket
process replies with {error, {notfound, toplevel_type()}}. Since the
purpose of modify_type/5 is to encapsulate a read-modify-write cycle,
I've added the 'create' option that will permit the not-found to be
treated as a way to create a new key by modifying an empty datatype.
If the option is not specified, the not-found error will be returned
normally.