change keepalive setting to mean true when specified in Options
An Erlang developer expects the atoms in options to denote the `true`
boolean value. So, when `keepalive` is in the options we expect that
keepalive feature is turned on.

In theory this is a backward incompatible change but in practice I cannot
imagine anyone setting `[keepalive, ...]` and expecting `[{keepalive, false},
...]` which is the default anyway.